Split registry sync code into separate `legacy_sync` module
